+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2019-09-26
+
 ### Added
 - Support for indexing Page Reference fields.
 - Support for indexing non-field Page properties (id, name).